Overhaul: completed !setup
Now that I know redbear is multiple server, we need to make some changes to the core. This is the beginning; convert to using Config for storing server-specific roles and channels.

This commit implements a setup command to help you check if the initial setup is complete, and set the values if not.